+++ lib/src/utils/export_utils.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart' show Color;
 import 'package:syncfusion_flutter_pdf/pdf.dart';
 import 'package:syncfusion_flutter_xlsio/xlsio.dart' as xls;
 import 'package:om_data_grid/src/utils/file_viewer/file_viewer.dart';
+import 'package:om_data_grid/src/utils/isolate_helper.dart';
 import '../models/grid_column_model.dart';
 import '../models/datagrid_configuration.dart';
 
@@ -16,14 +17,9 @@ class OmGridExportHandler {
     String fileName = 'Grid_Export.xlsx',
   }) async {
     try {
-      final List<int> bytes;
-      if (kIsWeb) {
-        bytes = _generateExcelBytes(data, columns, configuration);
-      } else {
-        bytes = await Isolate.run(
-          () => _generateExcelBytes(data, columns, configuration),
-        );
-      }
+      final List<int> bytes = await IsolateHelper.run(
+        () => _generateExcelBytes(data, columns, configuration),
+      );
 
       await _saveAndOpenFile(
         bytes,