+++ lib/src/utils/file_viewer/file_viewer_desktop.dart
@@ -1,0 +1,10 @@
+import 'package:url_launcher/url_launcher.dart';
+
+class FileViewerDesktop {
+  static Future<void> open(String filePath) async {
+    final Uri uri = Uri.file(filePath);
+    if (await canLaunchUrl(uri)) {
+      await launchUrl(uri);
+    }
+  }
+}