+++ lib/src/utils/file_viewer/file_viewer_io.dart
@@ -1,12 +1,19 @@
 import 'dart:io';
-import 'package:open_file_plus/open_file_plus.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/foundation.dart';
+import 'file_viewer_mobile.dart' if (dart.library.io) 'file_viewer_mobile.dart'
+    as mobile;
+import 'file_viewer_desktop.dart'
+    if (dart.library.io) 'file_viewer_desktop.dart' as desktop;
 
 class FileViewer {
   static Future<void> open(String filePath) async {
-    await OpenFile.open(filePath);
+    if (Platform.isAndroid || Platform.isIOS) {
+      await mobile.FileViewerMobile.open(filePath);
+    } else {
+      await desktop.FileViewerDesktop.open(filePath);
+    }
   }
 
   static Future<void> downloadAndOpen(String url, String fileName) async {
@@ -17,7 +24,7 @@ class FileViewer {
         final filePath = '${directory.path}/$fileName';
         final file = File(filePath);
         await file.writeAsBytes(response.bodyBytes);
-        await OpenFile.open(filePath);
+        await open(filePath);
       }
     } catch (e) {
       debugPrint("Download error: $e");
@@ -30,6 +37,6 @@ class FileViewer {
     final filePath = '${directory.path}/$fileName';
     final file = File(filePath);
     await file.writeAsBytes(bytes, flush: true);
-    await OpenFile.open(filePath);
+    await open(filePath);
   }
 }